src/requirements.txt
Flask
Flask-Cors
psycopg2-binary
flask_sqlalchemy